backend/requirements.txt
# NeuroPosture AI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy==1.24.3
scikit-learn==1.4.0
pydantic==2.6.0
pydantic-settings==2.1.0
websockets==12.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
motor==3.3.2
python-dotenv==1.0.0
